requirements/develop.txt
-r project.txt

mock==1.0.1
django-extensions==1.5.1
ipython[notebook]==3.0.0
factory_boy==2.4.1
